backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-docx
spacy
httpx
python-dotenv
